Migrate `device_grid` plugin to new danger API (#5060)
* Migrate `device_grid` plugin to new danger API

* Add missing `require`

* Do not use Bundler anymore

* Make hound happy and move spec to new gem

* Update `device_grid` README

* Remove `device_grid` spec from `fastlane`

It has been moved over to the separate `device_grid` gem.

* Rename directory to `danger-device_grid`
