+++ danger-device_grid/Gemfile
@@ -1,0 +1,12 @@
+source 'https://rubygems.org'
+
+gemspec
+
+group :development do
+  gem 'danger'
+
+  gem 'rspec'
+  gem 'rubocop'
+  gem 'webmock'
+  gem 'pry'
+end
