+++ danger-device_grid/lib/danger_device_grid.rb
@@ -1,0 +1,1 @@
+require 'device_grid/gem_version'
