+++ danger-device_grid/lib/danger_plugin.rb
@@ -1,0 +1,1 @@
+require 'device_grid/plugin'
