+++ danger-device_grid/lib/device_grid/gem_version.rb
@@ -1,0 +1,3 @@
+module DeviceGrid
+  VERSION = "0.1.0".freeze
+end
