+++ danger-device_grid/spec/spec_helper.rb
@@ -1,0 +1,10 @@
+require 'pathname'
+ROOT = Pathname.new(File.expand_path('../../', __FILE__))
+$:.unshift((ROOT + 'lib').to_s)
+$:.unshift((ROOT + 'spec').to_s)
+
+require 'bundler/setup'
+require 'rspec'
+require 'danger'
+
+require 'danger_plugin'
